# controllers/jobs.js
const Job = require("../models/Job");
const { StatusCodes } = require("http-status-codes");
const { BadRequestError, NotFoundError } = require("../errors");

const getAllJobs = async (req, res) => {
	const jobs = await Job.find({ createdBy: req.user.userId }).sort(
		"createdAt"
	);

	res.status(StatusCodes.OK).json({ jobs, count: jobs.length });
};

const getJob = async (req, res) => {
	const {
		user: { userId },
		params: { id: jobId },
	} = req;
	const job = await Job.findOne({
		createdBy: userId,
		_id: jobId,
	}).sort("createdAt");
	if (!job) throw new NotFoundError(`No job with id ${jobId}.`);
	res.status(StatusCodes.OK).json({ job });
};

const createJob = async (req, res) => {
	req.body.createdBy = req.user.userId;
	const job = await Job.create(req.body);
	res.status(StatusCodes.CREATED).json({ job });
};

const deleteJob = async (req, res) => {
	const {
		user: { userId },
		params: { id: jobId },
	} = req;
	const job = await Job.findByIdAndRemove(jobId);
	if (!job) throw new NotFoundError(`No job with id ${jobId}.`);
	res.status(StatusCodes.OK).send();
};

const updateJob = async (req, res) => {
	const {
		body: { company, position, status },
		user: { userId },
		params: { id: jobId },
	} = req;
	if (company.trim() === "" || position.trim() === "")
		throw new BadRequestError(
			"Company or Position fields can not be empty."
		);
	const statusList = ["interview", "declined", "pending"];
	if (status && !statusList.includes(status.trim())) {
		throw new BadRequestError(
			"Status can only be interview, declined or pending."
		);
	}
	const job = await Job.findByIdAndUpdate(
		{ _id: jobId, createdBy: userId },
		req.body,
		{ new: true, runValidators: true }
	);
	if (!job) throw new NotFoundError(`No job with id ${jobId}.`);
	res.status(StatusCodes.OK).json({ job });
};

module.exports = {
	getAllJobs,
	getJob,
	createJob,
	deleteJob,
	updateJob,
};